Get rid of literal function objects
Previously the query parameter parser functions were created during
macro-expansion time. This results in literal function objects being
produced, which are not externalizable objects.

In practice, this means that the resulting code can't be written to
FASL when doing file compilation.

This change postpones the creation of function objects so that they
are created after macro-expansion time.

After the change, e.g. compilation from slime works well with SBCL.